==> notebooks/00_IndoorAgriculture_start.ipynb <==
***Please note:*** you must restart the kernel to view the map in this notebook. (Press the double arrow button in the top menu bar.) It may take a few minutes to load on Binder, as there is * quite a lot of data* behind it.

# What's the State of Indoor Agriculture in the US?

Having worked in food distribution for a decade, I probably spend more time thinking about the logistics of food than most people do. When I eat, it's hard for me *not to* imagine where my food came from and how it reached me. Still, it wasn't until the pandemic approached New York City that the thought of losing access to fresh food had ever seemed like a real possiblity. The combination of fear and free time is powerful.

I found that my interest in gardening, long dormant, had reawoken virtually overnight. The ability to grow a few vegetables at home provided a little bit of comfort. It was really the knowledge, though, that food was being grown on a commercial scale within walking distance of my home in this crowded city that was able to nudge my mind closer to ease. For years, I'd happily been paying a premium for local hydroponically grown greens based on their quality alone. Yet, as the pandemic gained momentum and the city locked down, suddenly, what had once been to me purely source of luxury - commercial indoor farms - became a symbol of hope. 

Sure the world was a mess, but we might just be able to dig ourselves out of it.

## Where, exactly, are these high tech farms located?

Once I became attuned, it seemed as if indoor farms began cropping up left and right. Having caught wind of my interest, friends, family, and news apps fueled a steady stream of stories about the controlled environment agriculture (CEA) industry. News of startups developing indoor farms and more established players expanding came to me daily. Apparently, the industry was quite a bit larger than I'd previously realized. Seeking to contribute to an understanding of the indoor farming industry and the role it might play in strengthening overall food security, the question of *where* exactly these high tech farms were located seemed like a good starting point.  

## What kinds of places become home to indoor farms?
In addition to geographic location, I wanted to more about the *kinds* of places that become home to commercial indoor farms. Because there are so many different ways to characterize a place, I decided to narrow my initial scope to population density using data from the most recent US Census, 2020. Given that my own neighborhood can vary significantly in the span of a block or two, I wanted to work with the most granular population and geospatial data I could get. The *census block* is the smallest geographic unit for which census data is gathered, but when I first began this project, 2020 data at this level had not yet been released. Data at the *census tract* level, however, was available and was still relatively granular. I decided it would suffice and set out on my data journey.

## How do we tame the data beast?

While the *United States Decennial Census of Population and Housing* provides a very rich source of raw data, it can be challenging to decipher and synthesize into usable features. My notes walk through the data cleaning and analysis techniques I applied towards my goal of better understanding the indoor farming industry and its potential impact on food security. The "Data Sources" section of each notebook details the individual files used and the specific steps I followed to access them.

#### *Some helper code for managing file paths*

We'll use the code below throughout this project to make it easier to refer to the folders where our various data files are stored.

# the code below uses the `os` and `pathlib` modules to make it easier to
# refer to the various folders where data files for this project are stored
# we'll use it in each of the notebooks for this project

import os, pathlib
base_dir = pathlib.Path(os.getcwd()).parent
data_archive_dir = os.path.join(base_dir, "data_archive")
clean_data_dir = os.path.join(data_archive_dir, "clean")
data_dir = os.path.join(base_dir, "data")
shapes_dir = os.path.join(data_dir,"shapes")
json_dir = os.path.join(data_dir,"geojson")
util_dir = os.path.join(data_dir,"util")

## The 50,000 foot view: NYC population density

import pandas as pd

# load GEO DataFrame of the "place" New York City from CSV file
clean_geodf_place_nyc_file_csv  = os.path.join(clean_data_dir, "geodf_place_nyc.csv")
geodf_place_nyc = pd.read_csv(clean_geodf_place_nyc_file_csv, index_col=0)

# load GEO DataFrame of NY State census tracts from CSV file
clean_geodf_tract_ny_file_csv = os.path.join(clean_data_dir,'tl_2020_36_tract20.csv')
geodf_tract_ny = pd.read_csv(clean_geodf_tract_ny_file_csv, index_col=0)

from keplergl import KeplerGl

ny_map = KeplerGl(height=600, show_docs=False)
ny_map.add_data(data=geodf_place_nyc, name='New York City') 
ny_map.add_data(data=geodf_tract_ny, name= 'Population Density')

ny_map.config = {
    'version': 'v1',
 'config': {'visState': {'filters': [],
   'layers': [{'id': 'hjl4aki',
     'type': 'geojson',
     'config': {'dataId': 'New York City',
      'label': 'New York City',
      'color': [173, 185, 51],
      'highlightColor': [252, 242, 26, 255],
      'columns': {'geojson': 'geometry'},
      'isVisible': True,
      'visConfig': {'opacity': 0,
       'strokeOpacity': 0.8,
       'thickness': 1.0,
       'strokeColor': [173, 185, 51],
       'colorRange': {'name': 'Global Warming',
        'type': 'sequential',
        'category': 'Uber',
        'colors': ['#5A1846',
         '#900C3F',
         '#C70039',
         '#E3611C',
         '#F1920E',
         '#FFC300']},
       'strokeColorRange': {'name': 'Global Warming',
        'type': 'sequential',
        'category': 'Uber',
        'colors': ['#5A1846',
         '#900C3F',
         '#C70039',
         '#E3611C',
         '#F1920E',
         '#FFC300']},
       'radius': 10,
       'sizeRange': [0, 10],
       'radiusRange': [0, 50],
       'heightRange': [0, 500],
       'elevationScale': 5,
       'enableElevationZoomFactor': True,
       'stroked': True,
       'filled': True,
       'enable3d': False,
       'wireframe': False},
      'hidden': False,
      'textLabel': [{'field': None,
        'color': [255, 255, 255],
        'size': 18,
        'offset': [0, 0],
        'anchor': 'start',
        'alignment': 'center'}]},
     'visualChannels': {'colorField': None,
      'colorScale': 'quantile',
      'strokeColorField': None,
      'strokeColorScale': 'quantile',
      'sizeField': None,
      'sizeScale': 'linear',
      'heightField': None,
      'heightScale': 'linear',
      'radiusField': None,
      'radiusScale': 'linear'}},
    {'id': 'ge5p5d',
     'type': 'geojson',
     'config': {'dataId': 'Population Density',
      'label': 'Population Density',
      'color': [255, 203, 153],
      'highlightColor': [252, 242, 26, 255],
      'columns': {'geojson': 'geometry'},
      'isVisible': True,
      'visConfig': {'opacity': 0.8,
       'strokeOpacity': 0.8,
       'thickness': 0,
       'strokeColor': [248, 149, 112],
       'colorRange': {'name': 'Global Warming',
        'type': 'sequential',
        'category': 'Uber',
        'colors': ['#5A1846',
         '#900C3F',
         '#C70039',
         '#E3611C',
         '#F1920E',
         '#FFC300']},
       'strokeColorRange': {'name': 'Global Warming',
        'type': 'sequential',
        'category': 'Uber',
        'colors': ['#5A1846',
         '#900C3F',
         '#C70039',
         '#E3611C',
         '#F1920E',
         '#FFC300']},
       'radius': 10,
       'sizeRange': [0, 10],
       'radiusRange': [0, 50],
       'heightRange': [0, 500],
       'elevationScale': 5,
       'enableElevationZoomFactor': True,
       'stroked': True,
       'filled': True,
       'enable3d': False,
       'wireframe': False},
      'hidden': False,
      'textLabel': [{'field': None,
        'color': [255, 255, 255],
        'size': 18,
        'offset': [0, 0],
        'anchor': 'start',
        'alignment': 'center'}]},
     'visualChannels': {'colorField': {'name': 'Population Density',
       'type': 'integer'},
      'colorScale': 'quantile',
      'strokeColorField': None,
      'strokeColorScale': 'quantile',
      'sizeField': None,
      'sizeScale': 'linear',
      'heightField': None,
      'heightScale': 'linear',
      'radiusField': None,
      'radiusScale': 'linear'}}],
   'interactionConfig': {'tooltip': {'fieldsToShow': {'New York City': [{'name': 'State',
        'format': None},
       {'name': 'State FIPS', 'format': None},
       {'name': 'Place', 'format': None},
       {'name': 'Place FIPS', 'format': None},
       {'name': 'County', 'format': None}],
      'Population Density': [{'name': 'County FIPS', 'format': None},
       {'name': 'Census Tract Name', 'format': None},
       {'name': 'GEOID Census Tract', 'format': None},
       {'name': 'Population', 'format': None},
       {'name': 'Land Area', 'format': None}]},
     'compareMode': False,
     'compareType': 'absolute',
     'enabled': True},
    'brush': {'size': 0.5, 'enabled': False},
    'geocoder': {'enabled': False},
    'coordinate': {'enabled': False}},
   'layerBlending': 'normal',
   'splitMaps': [],
   'animationConfig': {'currentTime': None, 'speed': 1}},
  'mapState': {'bearing': 0,
   'dragRotate': False,
   'latitude': 40.70679298375131,
   'longitude': -74.0709215373596,
   'pitch': 0,
   'zoom': 9,
   'isSplit': False},
  'mapStyle': {'styleType': 'dark',
   'topLayerGroups': {},
   'visibleLayerGroups': {'label': True,
    'road': True,
    'border': False,
    'building': True,
    'water': True,
    'land': True,
    '3d building': False},
   'threeDBuildingColor': [9.665468314072013,
    17.18305478057247,
    31.1442867897876],
   'mapStyles': {}}}}

ny_map

## Up close: data cleaning and mapping

The following project notebooks provide a detailed look at the various steps of cleaning and mapping the data for this project. The final steps of creating a SQL database of indoor farms in New York City and then mapping them will follow shortly.

* [Data Cleanup: US Census population data](01_DataCleanup_Population.ipynb)
* [Data Cleanup: GEOID tables](02_DataCleanup_GEOIDs.ipynb)
* [Data Cleanup: Geographic data](03_DataCleanup_GIS.ipynb)
* [Mapping NY State: Population Density](04_Mapping_Population.ipynb)
* SQL Database Creation: Indoor Farms - *coming soon...*
* Mapping NY State: Indoor Farms - *coming soon...*
