==> notebooks/03_DataCleanup_GIS.ipynb <==
# Geographic data: Translating shapefiles to GEO DataFrames

The Census Bureau provides geographic information in *shapefile* format. In this notebook, we'll turn the raw shapefiles for NY State census tracts and "places" we downloaded earlier into clean, ready-to-map GEO DataFrames. We'll incorporate population density into our NY State census tract GEO DataFrame, save a separate GEO DataFrame for NYC, then export our GEO DataFrames for use in our next notebook on mapping.

These shapefiles are *huge* and there are a lot of them, so I don't put them in Github. Therefore, this notebook won't run in binder. but the code is here for viewing. I've commented out the code to save GEO DataFrames as Parquet files (because Arrow format for shape data is not final, so may cause problems. I plan to come back to this...)

#### *Data sources*

Files downloaded from Census Bureau FTP server via FTP client, file locations provided below.

* 2020 Census Redistricting Data (P.L. 94-171) Shapefiles, census tract level, NY State
  - file location: ftp2.census.gov/geo/tiger/TIGER2020PL/LAYER/TRACT/2020
  - file names beginning with "tl_2020_36" (NY State FIPS code is 36)

* 2020 Census Redistricting Data (P.L. 94-171) Shapefiles, places level, NY State
  - file location: ftp2.census.gov/geo/tiger/TIGER2020PL/LAYER/PLACE/2020
  - file name: tl_2020_36_place20.zip  (36 for NY State)

#### *Up close: data cleaning and mapping*

The notebooks below provide a detailed look at the other stages of cleaning and mapping the data for this project. The final steps of creating a SQL database of indoor farms in New York City and then mapping them will follow shortly.

* [Introduction: The State of Indoor Farming in the US] (00_IndoorAgriculture_start.ipynb) 
* [Data Cleanup: US Census population data](01_DataCleanup_Population.ipynb)
* [Data Cleanup: GEOID tables](02_DataCleanup_GEOIDs.ipynb)
<!-- * [Data Cleanup: Geographic data](03_DataCleanup_GIS.ipynb) -->
* [Mapping NY State: Population Density](04_Mapping_Population.ipynb)
* SQL Database Creation: Indoor Farms - *coming soon...*
* Mapping NY State: Indoor Farms - *coming soon...*

#### *Some helper code for managing file paths*
We'll use the code below throughout this project to make it easier to refer to the folders where our various data files are stored.

# os and patlib modules used to make it easier to refer to project folders 

import os, pathlib
base_dir = pathlib.Path(os.getcwd()).parent
data_archive_dir = os.path.join(base_dir, "data_archive")
clean_data_dir = os.path.join(data_archive_dir, "clean")
data_dir = os.path.join(base_dir, "data")
shapes_dir = os.path.join(data_dir,"shapes")
json_dir = os.path.join(data_dir,"geojson")
util_dir = os.path.join(data_dir,"util")

## Census tract shapefiles to GEO DataFrame

 Let's load our NY State census tract shapefiles as a GEO DataFrame and call it *geodf_tract_ny*.

import geopandas as gpd

big_tract_ny_shapefile = os.path.join(shapes_dir,"tl_2020_36_tract20.zip") # provide the full path to our shapefiles

# uses the geopandas function read_file to grab our file

# geodf_tract_ny = gpd.read_file(big_tract_ny_shapefile)[['STATEFP20', 'COUNTYFP20', 'TRACTCE20', 'GEOID20', 'ALAND20',  'AWATER20', 'INTPTLAT20', 'INTPTLON20', 'geometry']] 
geodf_tract_ny = gpd.read_file(big_tract_ny_shapefile)[['STATEFP20', 'COUNTYFP20', 'TRACTCE20', 'GEOID20', 'ALAND20',  'AWATER20', 'geometry']] 
# geodf_tract_ny = gpd.read_file(big_tract_ny_shapefile)
# import pyarrow as pa
# import warnings; 
# warnings.filterwarnings('ignore', message='.*initial implementation of Parquet.*')
# geodf_tract_ny.to_parquet(small_clean_shapefile, index=False, compression='BROTLI')
#  #= pa.Table.from_pandas(ny_shapes_df, preserve_index=False)

# rename the columns
# geodf_tract_ny = geodf_tract_ny.rename(columns={'STATEFP20': 'State FIPS', 'COUNTYFP20': 'County FIPS', 'TRACTCE20': 'Census Tract Code', 'GEOID20': 'GEOID Census Tract', 'ALAND20': 'Land Area', 'AWATER20': 'Water Area', 'INTPTLAT20': 'Latitude of Internal Point', 'INTPTLON20': 'Longitude of Internal Point'})
geodf_tract_ny = geodf_tract_ny.rename(columns={'STATEFP20': 'State FIPS', 'COUNTYFP20': 'County FIPS', 'TRACTCE20': 'Census Tract Code', 'GEOID20': 'GEOID Census Tract', 'ALAND20': 'Land Area', 'AWATER20': 'Water Area'})

# set the data types of each column as we want them to be
geodf_tract_ny = geodf_tract_ny.astype({'State FIPS': 'int', 'County FIPS':'int', 'Census Tract Code':'int', 'GEOID Census Tract': 'int', 'Land Area': 'int', 'Water Area': 'int'})

# NY State had 5411 census tracts total for the 2020 census (https://www.census.gov/geographies/reference-files/time-series/geo/tallies.html#tract_bg_block)
# so we want to see .shape to return 5411 rows below
geodf_tract_ny.shape 

geodf_tract_ny.head(3)

As usual, we'll conduct our checks for correct data types and missing values.

geodf_tract_ny.info()
geodf_tract_ny['geometry'].dtype

# double check for null values
geodf_tract_ny[geodf_tract_ny.isna().any(axis=1)]

No null values! We're making progress - our big NY Census Tract shapefile is now a nice *clean* GEO DataFrame.

### Merging population data and geographic data

We're now ready to use the population data we so carefully cleaned in our first notebook! Let's beging by loading as a pandas DataFrame the clean population data we saved earlier to parquet format. 

import pyarrow.parquet as pq

clean_census_pop_file_parquet = os.path.join(clean_data_dir, 'census_pop.parquet')
df_census_pop = pq.read_table(clean_census_pop_file_parquet, memory_map=True).to_pandas()
df_census_pop.head(3)

Now we can go ahead and merge *df_census_pop* into *geodf_tract_ny* on column "GEOID Census Tract". (We'll be calculating population density shortly, but first we need to check a few things...)

geodf_tract_ny = geodf_tract_ny.merge(df_census_pop, on=['GEOID Census Tract'])
geodf_tract_ny = geodf_tract_ny[['State FIPS', 'County FIPS', 'Census Tract Name', 'GEOID Census Tract', 'Population', 'Land Area', 'Water Area', 'geometry']]
geodf_tract_ny.head(3)

### Excluding census tracts that are water only

For reasons we plan to research further at some point, the US Census includes some census tracts that consist entirely of water and no land. It stands to reason that these would not have any permanent residents and would, therefore, have a recorded population of 0. If this is the case, we'll want to exclude them from our population density choropleth as they will create unecessary noise.

But, to be safe, let's test our assumption. First, let's see how many census tracts in NY state consist entirely of water.

geodf_tract_ny[geodf_tract_ny['Land Area']==0].count()

Apparently, there are a total of 18 census tracts in NY state than consist of water only! Now, let's see if we're correct that all of these have a population of 0.

geodf_tract_ny[(geodf_tract_ny['Land Area'] == 0) & (geodf_tract_ny['Population'] == 0)]

Confirmed - there are, in fact, 0 residents in any NY state census tract thats consist of water only. Now we can go ahead and exclude these census tracts from our Geo DataFrame. 

geodf_tract_ny = geodf_tract_ny[(geodf_tract_ny['Land Area'] != 0)]

# add a column to calculate population density
# land area measured in whole sq meters
# sq miles = sq meters/2,589,988 
# population density in sq miles  = population/ (land area sq meters / 2,589,988)
geodf_tract_ny['Population Density'] = (geodf_tract_ny['Population']/(geodf_tract_ny['Land Area']/2589988)).round(0)
geodf_tract_ny = geodf_tract_ny[['State FIPS', 'County FIPS', 'Census Tract Name', 'GEOID Census Tract', 'Population', 'Land Area', 'Water Area', 'Population Density', 'geometry']]
geodf_tract_ny.head(3)

Our population density calculation will have produced a NaN value in rows where "Population" == 0. So, let's locate these and replace each "NaN" in the "Population Density" column with a "0".

# tracts with population == 0 because these will result in NaN when we calculate Population Density
# so let's find these and replace the NaNs with 0s
geodf_tract_ny[geodf_tract_ny['Population'] == 0]
geodf_tract_ny['Population Density'] = geodf_tract_ny['Population Density'].fillna(0)
geodf_tract_ny['Population Density'] = geodf_tract_ny['Population Density'].astype(int) # change dtype to integer

Now let's take it a step further and also make our clean DataFrame *smaller* by saving it to Parquet format.

### Saving our clean GEO DataFrame for later

import pyarrow as pa
import warnings; 
warnings.filterwarnings('ignore', message='.*initial implementation of Parquet.*')
clean_geodf_tract_ny_file_parquet = os.path.join(clean_data_dir, "tl_2020_36_tract20.parquet")
geodf_tract_ny.to_parquet(clean_geodf_tract_ny_file_parquet, index=False, compression='BROTLI')
#= pa.Table.from_pandas(ny_shapes_df, preserve_index=False)

# we'll also save a CSV version 
clean_geodf_tract_ny_file_csv  = os.path.join(clean_data_dir, "tl_2020_36_tract20.csv")
geodf_tract_ny.to_csv(clean_geodf_tract_ny_file_csv, index=False)
#= pa.Table.from_pandas(ny_shapes_df, preserve_index=False)

## NY State "Place" shapefiles to GEO DataFrame

Let's load the NY State "Place" shapefiles as a GEO DataFrame and call it *geodf_place_ny*.

import geopandas as gpd 

# make GEO DataFrame from the NY State "Place" shapefiles
shapefile_place_ny = os.path.join(shapes_dir,"tl_2020_36_place20.zip")
geodf_place_ny = gpd.read_file(shapefile_place_ny, encoding_errors='ignore')

# select only the columns we want 
geodf_place_ny = geodf_place_ny[['STATEFP20', 'PLACEFP20', 'GEOID20', 'NAME20', 'ALAND20', 'AWATER20', 'geometry']]

# rename the columns
geodf_place_ny.rename(columns={'STATEFP20': 'State FIPS', 'PLACEFP20': 'Place FIPS', 'GEOID20': 'GEOID', 'NAME20': 'Name', 'ALAND20': 'Land Area', 'AWATER20': 'Water Area'}, inplace=True)

geodf_place_ny.head(3)
geodf_place_ny

Are there any "places" in NY state that consist of water only and no land?

geodf_place_ny[geodf_place_ny['Land Area'] == 0]

And, of course, we'll conduct our checks for missing values and correct datatypes.

geodf_place_ny.info()

The data type for State FIPS, Place FIPS, and GEOID should really be integer. Let's fix that now.

# Set the data type of each column are as we want it to be
geodf_place_ny['State FIPS'] = geodf_place_ny['State FIPS'].astype(int)
geodf_place_ny['Place FIPS'] = geodf_place_ny['Place FIPS'].astype(int)
geodf_place_ny['GEOID'] = geodf_place_ny['GEOID'].astype(int)
geodf_place_ny.dtypes

One final check for missing values.

# another check for null values
geodf_place_ny[geodf_place_ny.isna().any(axis=1)]

All clear! With null values rule out, we're almost done with our NY State place GEO DataFrame. There's just one last thing bothering us - a naming inconsistency between the NY place FIPS code list and NY place shapefiles we downloaded. This is confusing!

### Ensuring consistent naming

We can see an example below when we look at place FIPS 51000. It's referred to as "New York City" in the the NY Place FIPS code list and simply "New York" in the corresponding shapefiles. 

import pyarrow.parquet as pq

clean_df_place_file = os.path.join(clean_data_dir, 'place.parquet')
df_place = pq.read_table(clean_df_place_file, memory_map=True).to_pandas()
df_place[(df_place['Place FIPS'] == 51000) & (df_place['State FIPS'] == 36)]

geodf_place_ny[geodf_place_ny['Place FIPS'] == 51000]

We prefer the full name, "New York City". So, we'll resolve the inconsistency by merging the "Place" column from *df_place* into *geodf_place_ny* and keeping only the columns we want. 

geodf_place_ny = geodf_place_ny.merge(df_place, on =['State FIPS', 'Place FIPS'])
geodf_place_ny = geodf_place_ny[['State Name', 'State', 'State FIPS', 'Place', 'Place FIPS', 'County', 'Land Area', 'Water Area', 'geometry']] 
geodf_place_ny[geodf_place_ny['Place FIPS'] == 51000]

Good! We can now save this clean, consistent version as both a CSV and a Parquet file for use in our next notebook.

# save geodf_place_ny as Parquet
clean_geodf_place_ny_file_parquet = os.path.join(clean_data_dir, 'geodf_place_ny.parquet')
geodf_place_ny.to_parquet(clean_geodf_place_ny_file_parquet, index=False, compression='BROTLI')

# this is a back up CSV version 
clean_geodf_place_ny_file_csv  = os.path.join(clean_data_dir, "geodf_place_ny.csv")
geodf_place_ny.to_csv(clean_geodf_place_ny_file_csv, index=False)
#= pa.Table.from_pandas(ny_shapes_df, preserve_index=False)

## New York City "place" GEO DataFrame

Lastly, we'll save a separate GEO DataFrame of just the geographic information for New York City. We'll call it *geodfy_place_nyc* and also save it as a CSV and Parquet file. 

# make a new GEO DF of just the Place New York City
geodf_place_nyc = geodf_place_ny[geodf_place_ny['Place FIPS'] == 51000]
geodf_place_nyc 

# save as Parquet
clean_geodf_place_nyc_file_parquet = os.path.join(clean_data_dir, 'geodf_place_nyc.parquet')
geodf_place_nyc.to_parquet(clean_geodf_place_nyc_file_parquet, index=False, compression='BROTLI')

# also save a CSV version 
clean_geodf_place_nyc_file_csv = os.path.join(clean_data_dir, "geodf_place_nyc.csv")
geodf_place_nyc.to_csv(clean_geodf_place_nyc_file_csv, index=False)
#= pa.Table.from_pandas(ny_shapes_df, preserve_index=False)
